Put callbacks into utils() function
