+++ src/models/CNNDepthMap/CNNDepthMap-height/q3-depthmapmultiartifactlatefusion-plaincnn-height/src/utils.py
@@ -2,7 +2,9 @@ import datetime
 import os
 from pathlib import Path
 
+from azureml.core.run import Run
 from azureml.core.workspace import Workspace
+from tensorflow.keras import callbacks
 
 
 def download_dataset(workspace: Workspace, dataset_name: str, dataset_path: str):
@@ -17,3 +19,30 @@ def download_dataset(workspace: Workspace, dataset_name: str, dataset_path: str)
 
 def get_dataset_path(data_dir: Path, dataset_name: str):
     return str(data_dir / dataset_name)
+
+
+class AzureLogCallback(callbacks.Callback):
+    """Pushes metrics and losses into the run on AzureML"""
+    def __init__(self, run: Run):
+        super().__init__()
+        self.run = run
+
+    def on_epoch_end(self, epoch, logs=None):
+        if logs is not None:
+            for key, value in logs.items():
+                self.run.log(key, value)
+
+
+def create_tensorboard_callback() -> callbacks.TensorBoard:
+    return callbacks.TensorBoard(
+        log_dir="logs",
+        histogram_freq=0,
+        write_graph=True,
+        write_grads=False,
+        write_images=True,
+        embeddings_freq=0,
+        embeddings_layer_names=None,
+        embeddings_metadata=None,
+        embeddings_data=None,
+        update_freq="epoch"
+    )
